LayoutTests/http/wpt/fetch/response-opaque-clone.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Request clone</title>
    <meta name="help" href="https://fetch.spec.whatwg.org/#request">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script src="/common/get-host-info.sub.js"></script>
    <script>
promise_test(function(test) {
    return fetch(get_host_info().HTTP_REMOTE_ORIGIN, { mode: "no-cors" }).then((response) => {
        var clone = response.clone();
        assert_equals(response.body, null, "opaque response body should be null");
        assert_equals(clone.body, null, "clone body should be null");
        return clone.arrayBuffer();
    }).then((buffer) => {
        assert_equals(buffer.byteLength, 0, "cloned opaque response buffer should be null");
    });
}, "Check opaque response can be cloned");
    </script>
  </body>
</html>
